# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BuildPro Construction Business Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List all jobs
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: clientId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Create a new job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Created job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get a job by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Not found
    patch:
      operationId: updateJob
      tags: [jobs]
      summary: Update a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobUpdate"
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Delete a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /jobs/{id}/summary:
    get:
      operationId: getJobSummary
      tags: [jobs]
      summary: Get job financial and progress summary
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobSummary"

  # ──────────────────────────────────────
  # ESTIMATES
  # ──────────────────────────────────────
...
  /materials:
    get:
      operationId: listMaterials
      tags: [materials]
      summary: List materials with optional price comparison
      parameters:
        - name: zipCode
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of materials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Material"
    post:
      operationId: createMaterial
      tags: [materials]
      summary: Create a material in the catalog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaterialInput"
      responses:
        "201":
          description: Created material
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
  /materials/{id}:
    get:
      operationId: getMaterial
      tags: [materials]
      summary: Get a material by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Material
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
    patch:
      operationId: updateMaterial
      tags: [materials]
      summary: Update a material
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaterialUpdate"
      responses:
        "200":
          description: Updated material
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
    delete:
      operationId: deleteMaterial
      tags: [materials]
      summary: Delete a material
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /materials/{id}/prices:
    get:
      operationId: getMaterialPrices
      tags: [materials]
      summary: Get supplier price comparison for a material
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Price comparison list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SupplierPrice"
    post:
      operationId: addMaterialPrice
      tags: [materials]
      summary: Add a supplier price for a material
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierPriceInput"
      responses:
        "201":
          description: Created price entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierPrice"

  # ──────────────────────────────────────
  # JOB PHOTOS
  # ──────────────────────────────────────
...
  /events:
    get:
      operationId: listEvents
      tags: [calendar]
      summary: List calendar events
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: jobId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalendarEvent"
    post:
      operationId: createEvent
      tags: [calendar]
      summary: Create a calendar event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalendarEventInput"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarEvent"
  /events/{id}:
    get:
      operationId: getEvent
      tags: [calendar]
      summary: Get an event by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarEvent"
    patch:
      operationId: updateEvent
      tags: [calendar]
      summary: Update a calendar event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalendarEventUpdate"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarEvent"
    delete:
      operationId: deleteEvent
      tags: [calendar]
      summary: Delete a calendar event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    Job:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        clientId:
          type: ["integer", "null"]
        clientName:
          type: ["string", "null"]
        status:
          type: string
          enum: [lead, scheduled, in_progress, completed, cancelled]
        description:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        zipCode:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        estimatedValue:
          type: ["number", "null"]
        actualCost:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, title, status, createdAt]
    JobInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        clientId:
          type: integer
        status:
          type: string
          enum: [lead, scheduled, in_progress, completed, cancelled]
        description:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        estimatedValue:
          type: number
        notes:
          type: string
      required: [title]
    JobUpdate:
      type: object
      properties:
        title:
          type: string
        clientId:
          type: integer
        status:
          type: string
          enum: [lead, scheduled, in_progress, completed, cancelled]
        description:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        estimatedValue:
          type: number
        actualCost:
          type: number
        notes:
          type: string
    JobSummary:
      type: object
      properties:
        jobId:
          type: integer
        estimateTotal:
          type: number
        receiptTotal:
          type: number
        photoCount:
          type: integer
        beforePhotoCount:
          type: integer
        afterPhotoCount:
          type: integer
        estimateCount:
          type: integer
        acceptedEstimateTotal:
          type: number
      required:
        - jobId
        - estimateTotal
        - receiptTotal
        - photoCount
        - beforePhotoCount
        - afterPhotoCount
        - estimateCount
        - acceptedEstimateTotal

    # ── ESTIMATES ──
...
    Material:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
        unit:
          type: string
        basePrice:
          type: number
        sku:
          type: ["string", "null"]
        lowesUrl:
          type: ["string", "null"]
        homeDepotUrl:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, name, category, unit, basePrice, createdAt]
    MaterialInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        category:
          type: string
        unit:
          type: string
        basePrice:
          type: number
        sku:
          type: string
        lowesUrl:
          type: string
        homeDepotUrl:
          type: string
      required: [name, category, unit, basePrice]
    MaterialUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        unit:
          type: string
        basePrice:
          type: number
        sku:
          type: string
        lowesUrl:
          type: string
        homeDepotUrl:
          type: string
    SupplierPrice:
      type: object
      properties:
        id:
          type: integer
        materialId:
          type: integer
        supplierName:
          type: string
        price:
          type: number
        unit:
          type: string
        url:
          type: ["string", "null"]
        zipCode:
          type: ["string", "null"]
        inStock:
          type: boolean
        updatedAt:
          type: string
      required: [id, materialId, supplierName, price, unit, inStock, updatedAt]
    SupplierPriceInput:
      type: object
      properties:
        supplierName:
          type: string
          minLength: 1
        price:
          type: number
        unit:
          type: string
        url:
          type: string
        zipCode:
          type: string
        inStock:
          type: boolean
      required: [supplierName, price, unit]

    # ── JOB PHOTOS ──
...
    CalendarEvent:
      type: object
      properties:
        id:
          type: integer
        jobId:
          type: ["integer", "null"]
        jobTitle:
          type: ["string", "null"]
        clientId:
          type: ["integer", "null"]
        clientName:
          type: ["string", "null"]
        title:
          type: string
        type:
          type: string
          enum: [estimate_visit, job_start, job_work, job_end, follow_up, meeting, other]
        startDatetime:
          type: string
        endDatetime:
          type: string
        allDay:
          type: boolean
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, title, type, startDatetime, endDatetime, allDay, createdAt]
    CalendarEventInput:
      type: object
      properties:
        jobId:
          type: integer
        clientId:
          type: integer
        title:
          type: string
          minLength: 1
        type:
          type: string
          enum: [estimate_visit, job_start, job_work, job_end, follow_up, meeting, other]
        startDatetime:
          type: string
        endDatetime:
          type: string
        allDay:
          type: boolean
        notes:
          type: string
      required: [title, type, startDatetime, endDatetime]
    CalendarEventUpdate:
      type: object
      properties:
        jobId:
          type: integer
        clientId:
          type: integer
        title:
          type: string
        type:
          type: string
          enum: [estimate_visit, job_start, job_work, job_end, follow_up, meeting, other]
        startDatetime:
          type: string
        endDatetime:
          type: string
        allDay:
          type: boolean
        notes:
          type: string