# lib/api-zod/src/generated/types/jobSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BuildPro Construction Business Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface JobSummary {
  jobId: number;
  estimateTotal: number;
  receiptTotal: number;
  photoCount: number;
  beforePhotoCount: number;
  afterPhotoCount: number;
  estimateCount: number;
  acceptedEstimateTotal: number;
}